requirements.txt
PySide2==5.12.3
mysql-connector-python==8.0.16
numpy==1.16.4
pandas==0.24.2
protobuf==3.8.0
python-dateutil==2.8.0
pytz==2019.1
shiboken2==5.12.3
six==1.12.0
xlrd==1.2.0
